Fix compilation errors: Add missing Theme/Components and fix SwiftData usage
- Added ColorTheme.swift with color extensions (backgroundPrimary, espressoBrown, etc.)
- Added CustomCard.swift and CustomButton.swift components
- Added Info.plist with app configuration
- Fixed @ObservedObject usage: SwiftData models don't need @ObservedObject wrapper
- Fixed Int16 to Int for Grinder.burrSize to match SwiftData model
- Removed @ObservedObject from: Bean, Grinder, Machine, BrewingSession properties
- Updated 7 view files to properly handle SwiftData models

SwiftData models marked with @Model are automatically observable and should
be used directly without property wrappers.

+++ EspressoTracker/Views/Equipment/AddGrinderView.swift
@@ -119,7 +119,7 @@ struct AddGrinderView: View {
     }
 
     private func saveGrinder() {
-        let burrSizeValue = Int16(burrSize) ?? 0
+        let burrSizeValue = Int(burrSize) ?? 0
         let dataManager = DataManager(modelContext: modelContext)
         dataManager.createGrinder(
             name: name,
